CBSOData4.jl

Documentation for CBSOData4.jl

CBSOData4.ODataTable_long — Method

ODataTable_long(table; columns, filters, base, catalog)

Get a Tables object for the dataset defined by table. Optional parameters are:

columns::Vector{String}, exact columnnames to select, empty for all columns

filters::String, OData4 specification for row filter

base::String, basename of the OData4 server

catalog::String, path part of the OData4 server for catalog information

Examples

tbl = CBSOData4.ODataTable_long("82931NED", filter="Measure in ['T001036']")

CBSOData4.get_catalogs — Function

get_catalogs(base)

Get all the catalogs as a list of dicts. The Identifier can be used as the catalog for queries about datasets in a certain catalog. If no parameters are given, the data is retrieved from the (CBS)[https://www.cbs.nl] OData4 portal.

The optional parameter is:

base::String, basename of the OData4 server

CBSOData4.get_meta — Method

get_meta(table; base, catalog)

Get the metadata for a given `table`. The result is a dict with the keys `"Dimensions"`, `"MeasureCodes"`, `"Properties"` and all the classifications (seperate codes and groups).

Optional parameters are:

base::String, basename of the OData4 server

catalog::String, path part of the OData4 server for regular data

CBSOData4.get_tables — Function

get_tables(base, catalog)

Get all the tables in the catalog as a list of dicts. The Identifier can be used to get the actual data. If no parameters are given, the data is retrieved from the (CBS)[https://www.cbs.nl] OData4 portal.

The optional parameters are:

base::String, basename of the OData4 server

catalog::String, path part of the OData4 server for catalog information